maj middle W49 email & auth track sandbox prod -try 1

+++ app/api/demande_collecte/delete_bsd/route.ts
@@ -1,17 +1,28 @@
 import { supabase } from "@/app/database/supabaseClient";
 import { NextResponse } from "next/server";
 import axios from "axios";
-
-const token_sandbox = 'tCJJTq0Da55LuoJMc35QEqwomMRDwl10xT1hI2UV';
-const url_sandbox = 'https://api.sandbox.trackdechets.beta.gouv.fr';
+import { cookies } from 'next/headers';
 
 export async function POST(request: Request) {
     const { id } = await request.json();
+    
+    const token_track = cookies().get('trackdechets_token')?.value;
+    let url_track = process.env.TRACKDECHETS_URL_SANDBOX;
+    if(process.env.NEXT_PUBLIC_TRACK_TYPE === 'app'){
+        url_track = process.env.TRACKDECHETS_URL_APP;
+    }
+    if (!token_track || !url_track) {
+        return NextResponse.json({ 
+            success: false, 
+            message: 'Token ou URL TrackDéchets non trouvés',
+            error: 'Token ou URL TrackDéchets non trouvés'
+        }, { status: 400 });
+    }
 
     //Check if the BSD is on TrackDéchets
     const { data:bsd, error:bsdError } = await supabase.from('bsd').select('on_track_dechets, status_track_dechets, id_track_dechets').eq('id', id).single();
     if(bsd?.on_track_dechets){
-        const trackDechetsResponse = await deleteBSDD_API(bsd.id_track_dechets);
+        const trackDechetsResponse = await deleteBSDD_API(bsd.id_track_dechets, token_track, url_track);
         if(trackDechetsResponse.status === 200){
             //Delete from Supabase
             const { data, error } = await supabase.from('bsd').delete().eq('id', id);
@@ -22,7 +33,7 @@ export async function POST(request: Request) {
     return NextResponse.json({ data, error });
 }
 
-const deleteBSDD_API = async (id:string) => {
+const deleteBSDD_API = async (id:string, token_track:string, url_track:string) => {
 
     const query = `
     mutation {
@@ -32,11 +43,11 @@ const deleteBSDD_API = async (id:string) => {
     }
     `;
     const response = await axios.post(
-        url_sandbox,
+        url_track,
         { query: query },
         {
             headers: {
-                Authorization: `Bearer ${token_sandbox}`,
+                Authorization: `Bearer ${token_track}`,
                 'Content-Type': 'application/json'
             }
         }
